refactor: その他シーンをsrc/scenes/に移行（TASK-0096）
TitleScene, RankUpScene, ShopScene, GameOverScene, GameClearSceneを
src/presentation/scenes/からsrc/scenes/に移行し、
旧パスには再エクスポートシムを配置して後方互換性を維持。
scenes/index.tsのバレルエクスポートを更新。
