+++ atelier-guild-rank/src/scenes/index.ts
@@ -2,10 +2,16 @@
  * Scenes Module - Phaserシーン公開API
  *
  * TASK-0094: scenes/ディレクトリ作成とBootScene移行
+ * TASK-0096: その他シーン移行（RankUp, Shop, GameOver, GameClear, TitleScene）
  *
  * @description
  * Phaserシーンの公開エクスポート。
- * 各シーンは段階的にsrc/scenes/に移行される。
  */
 
 export { BootScene } from './BootScene';
+export { GameClearScene } from './GameClearScene';
+export { GameOverScene } from './GameOverScene';
+export { RankUpScene } from './RankUpScene';
+export { ShopScene } from './ShopScene';
+export type { ISaveDataRepository, SaveData } from './TitleScene';
+export { TitleScene } from './TitleScene';
